Feat(Day7): Enable Dual-Core, Add TDD RingBuffer & Refactor CMake

+++ src/app/app_blink.c
@@ -4,7 +4,7 @@
 #include "hal_led.h"
 #include "hal_multicore.h"
 
-// 修正 2: 加入 static 關鍵字，因為這個函式只在這個檔案內使用
+// Core 1 任務
 static void core1_worker_task(void)
 {
     uint32_t cmd = 0;
@@ -13,17 +13,18 @@ static void core1_worker_task(void)
     {
         if (hal_multicore_fifo_pop(&cmd))
         {
-            // 修正 1: 將 %lu 改為 %u
-            printf("[Core 1] Received command: %u\n", cmd);
+            // 【關鍵修正】強制轉型為 (unsigned long) 搭配 %lu
+            // 這能同時騙過靜態分析器，並滿足 ARM 編譯器
+            printf("[Core 1] Received command: %lu\n", (unsigned long)cmd);
 
             hal_multicore_fifo_push(cmd * 2);
         }
     }
 }
 
+// App 主程式
 void app_blink_run(void)
 {
-    // 取得 LED 裝置物件
     const LedDevice *led = hal_led_get_default();
 
     if (led && led->init)
@@ -44,15 +45,16 @@ void app_blink_run(void)
             led->toggle();
         }
 
-        // 修正 1: 將 %lu 改為 %u
-        printf("[Core 0] Blinking LED... Counter: %u\n", counter);
+        // 【關鍵修正】強制轉型
+        printf("[Core 0] Blinking LED... Counter: %lu\n", (unsigned long)counter);
 
         hal_multicore_fifo_push(counter);
 
         if (hal_multicore_fifo_pop(&result))
         {
-            // 修正 1: 將 %lu 改為 %u
-            printf("[Core 0] Core 1 replied: %u (Double of %u)\n", result, counter);
+            // 【關鍵修正】強制轉型
+            printf("[Core 0] Core 1 replied: %lu (Double of %lu)\n", (unsigned long)result,
+                   (unsigned long)counter);
         }
 
         counter++;
